+++ build_scripts/qt/sources.pri
@@ -13,7 +13,7 @@ SOURCES += src/main.cpp\
     src/utils/ChaperoneUtils.cpp \
     src/tabcontrollers/audiomanager/AudioManagerDummy.cpp \
     src/tabcontrollers/keyboardinput/KeyboardInputDummy.cpp \
-    src/overlaycontroller/openvr_init.cpp \
+    src/openvr/openvr_init.cpp \
     src/ivrinput/ivrinput.cpp
 
 HEADERS += src/overlaycontroller.h \
@@ -34,12 +34,13 @@ HEADERS += src/overlaycontroller.h \
     src/quaternion/quaternion.h \
     src/tabcontrollers/audiomanager/AudioManagerDummy.h \
     src/tabcontrollers/keyboardinput/KeyboardInputDummy.h \
-    src/overlaycontroller/openvr_init.h \
+    src/openvr/openvr_init.h \
     src/ivrinput/ivrinput_action.h \
     src/ivrinput/ivrinput_manifest.h \
     src/ivrinput/ivrinput_action_set.h \
     src/ivrinput/ivrinput.h \
     src/utils/setup.h
+
 
 win32 {
     SOURCES += src/tabcontrollers/audiomanager/AudioManagerWindows.cpp \
